GN: Add video_engine_tests
Adds separate source_sets for the video_engine_tests subtargets inside
audio, call and video and merges them together into video_engine_tests.

BUG=webrtc:5949
R=kjellander@webrtc.org
TBR=mflodman@webrtc.org

Review URL: https://codereview.webrtc.org/2064523002 .

Cr-Commit-Position: refs/heads/master@{#13127}

+++ webrtc/audio/audio_receive_stream_unittest.cc
@@ -187,7 +187,7 @@ void BuildOneByteExtension(std::vector<uint8_t>::iterator it,
   const size_t kExtensionDataLength = kOneByteExtensionLength - 1;
   uint32_t shifted_value = extension_value
                            << (8 * (kExtensionDataLength - value_length));
-  *it = (id << 4) + (value_length - 1);
+  *it = (id << 4) + (static_cast<uint8_t>(value_length) - 1);
   ++it;
   ByteWriter<uint32_t, kExtensionDataLength>::WriteBigEndian(&(*it),
                                                              shifted_value);
